feat step3 자동차 경주 게임 - 경기 결과 기록 목적만 되도록 기능 수정 isMove 메서드 이동
todo test 코드 작성
- Random 메서드 전략 패턴을 사용하여 테스트하기 어려운 메소드 stub 데이터로 반환하도록 수정
