+++ src/main/java/racing/core/RacingCarGame.java
@@ -6,10 +6,13 @@ import racing.vo.RacingCarGamePlayInfo;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Random;
 
 
 public class RacingCarGame {
 
+  public static final int LIMIT_RANDOM_NUMBER = 10;
+  public static final int MORE_THAN_NUMBER_FOR_CAR_MOVE = 4;
   private final RacingCarGamePlayInfo info;
 
   private RacingCarGame(RacingCarGamePlayInfo info) {
@@ -23,13 +26,24 @@ public class RacingCarGame {
   public List<RacingRound> play() {
     int totalRound = info.getTotalRound();
     List<RacingCar> racingCars = info.getPlayers();
-    List<RacingRound> result = new ArrayList<>(totalRound);
+
+    List<RacingRound> racingRounds = new ArrayList<>(totalRound);
     for (int i = 0; i < totalRound; i++) {
-      RacingRound round = new RacingRound();
-      round.record(racingCars);
-      result.add(round);
+      for (RacingCar car : racingCars) {
+        if (isMove())
+          car.move();
+      }
+      racingRounds.add(RacingRound.createRecord(racingCars));
     }
-    return result;
+    return racingRounds;
+  }
+
+  private boolean isMove() {
+    return getRandomNumber() >= MORE_THAN_NUMBER_FOR_CAR_MOVE;
+  }
+
+  private int getRandomNumber() {
+    return new Random().nextInt(LIMIT_RANDOM_NUMBER);
   }
 
 }
